+++ tree_test.go
@@ -267,7 +267,7 @@ func TestOffset2key8BitsWide(t *testing.T) {
 	key, _ := hex.DecodeString("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f")
 	for i := 0; i < 32; i++ {
 		childId := offset2key(key, i*8)
-		if childId != uint(i) {
+		if childId != byte(i) {
 			t.Fatalf("error getting child number in key %d != %d", childId, i)
 		}
 	}
